requirements.txt
azure-cli
httpie
parse
PyGithub
questionary
rich
todoist-python
